+++ common/src/main/java/org/alfresco/hxi_connector/common/model/BulkIngesterEvent.java
@@ -2,7 +2,7 @@
  * #%L
  * Alfresco HX Insight Connector
  * %%
- * Copyright (C) 2024 Alfresco Software Limited
+ * Copyright (C) 2023 - 2024 Alfresco Software Limited
  * %%
  * This file is part of the Alfresco software.
  * If the software was purchased under a paid Alfresco license, the terms of
@@ -24,17 +24,14 @@
  * #L%
  */
 
-package org.alfresco.hxi_connector.live_ingester.adapters.messaging.bulk_ingester.model;
+package org.alfresco.hxi_connector.common.model;
 
 import java.io.Serializable;
 import java.util.Map;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
-import org.springframework.validation.annotation.Validated;
-
-@Validated
-public record BulkIngesterEventt(
+public record BulkIngesterEvent(
         @NotBlank String nodeId,
         ContentInfo contentInfo,
         @NotNull Map<String, Serializable> properties)
@@ -42,6 +39,6 @@ public record BulkIngesterEventt(
     public record ContentInfo(
             long contentSize,
             String encoding,
-            String mimetype)
+            String mimetype) implements Serializable
     {}
 }
