+++ live-ingester/pom.xml
@@ -59,6 +59,11 @@
             <groupId>com.fasterxml.jackson.core</groupId>
             <artifactId>jackson-databind</artifactId>
         </dependency>
+
+        <dependency>
+            <groupId>org.alfresco</groupId>
+            <artifactId>alfresco-hxinsight-connector-common</artifactId>
+        </dependency>
         <dependency>
             <groupId>org.alfresco</groupId>
             <artifactId>acs-event-model</artifactId>
