+++ live-ingester/src/main/java/org/alfresco/hxi_connector/live_ingester/adapters/messaging/bulk_ingester/BulkIngesterEventProcessor.java
@@ -37,6 +37,7 @@ import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
+import org.springframework.validation.annotation.Validated;
 
 import org.alfresco.hxi_connector.live_ingester.adapters.messaging.bulk_ingester.model.BulkIngesterEvent;
 import org.alfresco.hxi_connector.live_ingester.domain.usecase.content.IngestContentCommand;
@@ -53,7 +54,7 @@ public class BulkIngesterEventProcessor
     private final IngestMetadataCommandHandler ingestMetadataCommandHandler;
     private final IngestContentCommandHandler ingestContentCommandHandler;
 
-    public void process(BulkIngesterEvent event)
+    public void process(@Validated BulkIngesterEvent event)
     {
         Map<String, Serializable> properties = event.properties().entrySet().stream()
                 .collect(Collectors.toMap(Entry::getKey, Entry::getValue));
